Separated merging response jsonls into condition json into its own Python script
